safe shrinking fix #481

+++ src/main/java/com/simibubi/create/compat/jei/category/CreateRecipeCategory.java
@@ -135,7 +135,7 @@ public abstract class CreateRecipeCategory<T extends Recipe<?>> implements IReci
 
 	public static FluidStack withImprovedVisibility(FluidStack stack) {
 		FluidStack display = stack.copy();
-		int displayedAmount = (int) (stack.getAmount() * .75f) + 250;
+		long displayedAmount = (long) (stack.getAmount() * .75f) + 250;
 		display.setAmount(displayedAmount);
 		return display;
 	}
